requirements .txt
python-telegram-bot==20.7
firebase-admin==6.4.0